+++ settingrepo-node/src/main/java/com/dwarfeng/settingrepo/node/launcher/Launcher.java
@@ -1,8 +1,8 @@
 package com.dwarfeng.settingrepo.node.launcher;
 
 import com.dwarfeng.settingrepo.node.handler.LauncherSettingHandler;
-import com.dwarfeng.settingrepo.stack.service.FormatterSupportMaintainService;
 import com.dwarfeng.settingrepo.stack.service.ResetQosService;
+import com.dwarfeng.settingrepo.stack.service.SupportQosService;
 import com.dwarfeng.springterminator.sdk.util.ApplicationUtil;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import org.slf4j.Logger;
@@ -40,15 +40,18 @@ public class Launcher {
         // 获取启动器设置处理器，用于获取启动器设置，并按照设置选择性执行功能。
         LauncherSettingHandler launcherSettingHandler = ctx.getBean(LauncherSettingHandler.class);
 
-        // 判断是否重置格式化器支持。
-        if (launcherSettingHandler.isResetFormatterSupport()) {
-            LOGGER.info("重置格式化器支持...");
-            FormatterSupportMaintainService maintainService = ctx.getBean(FormatterSupportMaintainService.class);
-            try {
-                maintainService.reset();
-            } catch (ServiceException e) {
-                LOGGER.warn("格式化器支持重置失败，异常信息如下", e);
-            }
+        // 如果不重置格式化器，则返回。
+        if (!launcherSettingHandler.isResetFormatterSupport()) {
+            return;
+        }
+
+        // 重置格式化器支持。
+        LOGGER.info("重置格式化器支持...");
+        SupportQosService supportQosService = ctx.getBean(SupportQosService.class);
+        try {
+            supportQosService.resetFormatter();
+        } catch (ServiceException e) {
+            LOGGER.warn("格式化器支持重置失败，异常信息如下", e);
         }
     }